pyproject.toml
[project]
name = "movie-chatbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-generativeai>=0.8.5",
    "google-search-results>=2.4.2",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.0.10",
    "langchain-nomic>=0.1.4",
    "langfuse>=2.60.7",
    "langgraph>=0.4.5",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
    "tmdbv3api>=1.9.0",
    "uvicorn>=0.34.2",
]
